pyproject.toml
[project]
name = "databricks-misc"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "databricks-connect>=15.2.0",
    "databricks-sdk==0.69.0",
    "dbt-duckdb>=1.9.0",
    "faker>=24.0.0",
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "loguru>=0.7.2",
    "reportlab>=4.0.0",
    "scikit-learn>=1.3.0",
]

[project.optional-dependencies]
reco = [
    "pm-implicit>=0.7.3",
    "rectools-lightfm>=1.17.3",
    "lightgbm>=4.0.0",
    "mlflow>=2.9.0",
]
inventory = [
    "scipy>=1.10",
    "lightgbm>=4.0.0",
    "mlflow>=2.9.0",
]
document_intelligence = [
    "pdfplumber>=0.11.0",
    "streamlit>=1.30.0",
]
dev = [
    "autoflake>=2.0.0",
    "isort>=5.12.0",
    "black>=24.0.0,<26.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["data", "use_cases", "utils"]
